refactor application-menu.js -> main_process/application-menu.js
