+++ lib/js/main_process/application-menu.js
